requirements.txt
python-binance==1.0.19
python-dotenv==1.0.1
colorama==0.4.6
tabulate==0.9.0
